New hardware design for start module
Performed changes:
- Redesign of board
- Changed the microcontroller from Attiny13A to Attiny24/44A
- Changed the ir-receiver from TSOP75x to TSOP85x_AP5

Resons for these changes are mainly the following two points:
- 1. The IR-receiver used in the previous version was hard to get ahold
of in the correct packaging used for pick-and-place.
- 2. The memory of the microcontroller was a little to small, which
menas the code became very hard to read. Changing to one with more
memory, and more pins.

One thing to discuss if more components should be added, such as an
extra LED.