Some corrections from 1.0
Due to some errors, I decided to switch to 0402 components, thus making
it easier to route and to add another feature: a serial resistor for
the home-pad. The signals for the pads used for the pogo-bead are
changed. The positions are still the same, but the signals are a little
mixed.