Version 0.17
Since it was not very expensive to get the manufacturer to flip the
IR-receiver, it's best to go with a version similar to the old one
(V0.16). What is new from that verison is a new microcontroller
(Attiny44A), anothe LED and resistors in 0402.
Furthermore, the version with another IR-receiver (V1.0) has been
slightly updated.